requirements.txt
chromadb==0.5.7
cryptography==43.0.3
email-validator==2.2.0
fastapi-cli==0.0.5
h2==4.1.0
httptools==0.6.1
jaraco.collections==5.1.0
jinja2==3.1.4
langchain-community==0.2.18
langchain-openai==0.1.25
langchain-pinecone==0.1.3
lorem-text==2.1
pinecone==5.3.1
pip-chill==1.0.3
pipupgrade==1.12.0
platformdirs==4.2.2
pydantic-settings==2.5.2
pyjwt==2.9.0
python-jose==3.3.0
python-multipart==0.0.10
supabase==2.9.1
uvicorn==0.32.0
ujson==5.10.0
uvloop==0.20.0
watchfiles==0.24.0
